Initial implementation of rendering chain
A lot of changes here:

First, the factory function also comes with some methods. This will be
useful for exposing other ways of creating instances.

Second, the method for handling asynchronous rendering is to return
an Observable from the `render` function. We chain the streams
together to handle the asynchronous rendering pipeline, updating via
HTML-returning template function then a post-render stream.

The streams are even cancelable, so if another render takes place, the
previous rendering pipeline stops. This allows us to write animations,
without worrying about them interfering with each other.

Comes with a few tests, but a lot of these functions need to be
extracted and tested independantly.

Updates kefir to a commit hash version to use new `observer` method.
Also needed to update the aliases for `redux` and `kefir` to point
to their es6 versions.