Replace `makeEventSwapper` with map impl
Maps the `render$` stream to a new stream that, at end, emits a new
`events$` stream. That `events$` stream is bootstrapped with the
initial `events$` stream and `flatMap`'d to the final `events$` stream
returned from the function. This is how we turn a series of `render$`
streams into an updating `events$` stream.