Rewrite delegated DOM event system
This makes the process simpler: All the DOM events pass through a
single stream. When that stream is subscribed to, those events are
delegated on the body, so there's no need to manage whether the
listeners are attache or not. They're only attached if something is
subscribing to them. This wraps all of our delegating logic into whether
the stream is subscribed to or not.

All of the `events$` streams are just mapped & filtered from the
`sources$` stream.