Rename enhancer to observeDelta and add README
A river delta is a landform that forms from deposition of sediment carried
by a river as the flow leaves its mouth and enters slower-moving or
standing water. This occurs where a river enters an ocean, sea, estuary,
lake, reservoir, or (more rarely) another river that cannot transport away
the supplied sediment. - Wikipedia

This makes for a use metaphor, because a delta `source$` is a stream
that gets merged into a larger "body" (the `delta$` stream) and
"deposits" its actions into it.

Ok, it's not pefect, but I still like it.