Extract events$ stream generator function
Similar to the render$ stream generator, this function takes a
configuration object and returns an event stream generator.
This can easily be generated seperately from the component
definition and tested on its own. Provide the returned function
and element, fire events on it, and ensure the returned value is
what you expect it to be.