Extract helpers into module & combine w/ utils
`helpers` was originally designed to be Handlebars helpers, but
it makes more sense to provide them as a set of helper functions,
which developers can pull in a register to Handlebars themselves.
This way, they have the flexibility to decide how the helper works
by writing their own wrapper functions.