Cleanup component implementation
The component is really just a combination of 3 streams. If we
create 3 streams which, using the provided functions, emit
streams as values themselves, we can `flatMapLatest` the entire
`instance$` stream at once. We then merge in the `onMount`
stream, and ensure the whole stream only emits on actions.

Also updates some references from `state$` ->`props$`. For the
nomeclature of `brookjs`, `state$` is a stream of `state` objects,
which is the single application state atom, managed by Redux.
`props$` is a stream of `props` objects, which is the single UI
props atom, selected from the application state and passed down
to the components. Currently, `subcomponents` supports passing
an `adapter`, which selects the subcomponent's `props` from the
props passed down by its parent. That mechansim made
debugging diffcut by spreading around the areas where bugs can
occur, rather than centralizing them in a single selector.

The next version of the `subcomponents`-like implementation
will encourage the `state` to be mapped to `props` in a way that
matches the overall component hierarchy.