Rename adapter -> modifyChildProps & add preplug
`modifyChildProps` is more descriptive of what the function does,
receiving the full `props$` stream and returning a modified instance
of the `props$` that will be passed to the child's factory.

`preplug` is passed the final `instance$` before it's plugged into
the `children$` stream, providing a hook to modify the `instance$`.