Documentation & cleanup of children
Adds some extra inline documentation to explain what's going on.
Some work needs to be done to iron out a bug when using the
combinator in the `component`. If you use the mixin streams, the
`mutation$` streams won't get subscribed to and the `children$`
won't get new nodes. Each individual stream needs to be made by
subscribing to the corresponding `mutations$` stream.