Extract util functions & refactor implementation
Instead of returning the pool directly, we wrap the returning of the
pool in a merged$ stream that `flatMapLatest` to the `pool$`
itself. This allows subscribing to the `mutations$` to be turned on
and off w/ the subscribing and unsubcribing of the `pool$`.
Everything gets tied together in one stream.