Eliminate unnecessary `flatMapLatest`
The problem will the implementation as currently designed is that
it's constantly subscribing and unsubscribing to streams, even
if those streams are the same streams that came before, like the
constant(children$). The means that all of the startup & teardown
logic was running over and over again, which was very heavy.

This problem was eliminated in two ways:

* In the children stream generator, the side effect of plug/unplug
remains tucked away in the `diff` method, but we merge it into
the primary stream anyway so the streams get activated, running
the side effect as new elements get added and removed from the
DOM.

* In the component generator, the combinating of the streams
happens once, moving those `flatMapLatest` up to the streams
themselves rather than after the streams have been combined.
The means that combinator only executes once.