Don't set modifyChildProps to grab prop(key)
This implicit behavior is, while possibly convenient, very confusing when
setting things up. If you don't _need_ selectors to map the state to the
props, given the flexibility you may have with e.g. a PHP backend, and
it can be difficult to debug that behavior when there's nothing in
userland code that indicates what's happening.

The framework needs to be explicit. Implicit behavior like this is hard
to reason about.