Replace child node if the container is to change
This indicates that a children component would be mutated into a new
component. Morphdom normally updates the tree, rather than discarding
the node and creating a new one (which is what React does). Catching
these instances and doing a full discard/replace is computationally
cheaper than diffing the entire subtree.