Convert observeDelta from Enhancer to Middleware
As much as the Enhancer was a clean & pretty solution, it does not
play well with Middleware. Because the `dispatch` function is
wrapped by each enhancer like a series of concentric circles, actions
from the inner Enhancer were unable to pass through to the outer
Enhancer.

Middleware wraps the `store`'s `dispatch` function and passes the
result to the next function in the chain, which then proceed to wrap
it again. The issue is that the Middleware is unable to get access to
the newly wrapped `dispatch`, so the actions dispatched by the
Delta don't get dispatched through the Middleware.

These means, depending on the order the enhancers are composed,
one is unable to "speak" to the other, as the actions from one are
unable to flow to the other, which makes things extremely difficult
to code. Converting this to a middleware, rather than going the
enhancer route, ensures any middleware can communicate with any
of the deltas.