Add pre-deprecated props fallback
If the container attribute is passed in via the hbs template, the value
may not be in the `props$` stream passed into the templated
component. This means, when rendering the Handlebars template,
the container attribute may not have the value set, which would
cause morphdom to set it to a string, messing up the children
bindings.

For now, we're going to ensure that anytime the attribute would be
morphed into an empty string, we reset it to the original and
issue a deprecation notice, allowing the developer time to update
the `props$` streams accordingly.