Fix events tests in PhantomJS
`cut` & `paste` do not work in PhantomJS because there is no
`ClipboardEvent` in that environment. Which makes sense, because
it's headless, but it does mean the tests don't pass there. The tests
do pass in Chrome, which has a `ClipboardEvent`, so we know it works,
but this ensures we don't get failing tests in PhantomJS.