Add `key` configuration for iterable children
This ensures a component with multiple instances gets its props
distributed properly. Currently, this does require configuration both in
the Handlebars template as well as the JS, but does make sure that any
subcomponent gets the correct props, as well as ensure `morphdom`
will merely rearrange subcomponents with the same key.