Ensure subnodes are mounted correctly
Previously if the node added, as recorded in the MutationRecord,
didn't have a container attribute, it wasn't transmitted down the stream.
This excluded child elements that were mounted within the added node.
This checks if there are any top-level children mounted to the added
node and ensures they're dispatched down the mutation stream.