Update render api to take el & props$ and add docs
Setting up the render$ stream to respond to `props$` & `raf$` pulls
control of the render cycle out of the `component` and into the
`render` module, which we can use to provide the `render` module
with more control over the render cycle and expose that control to
userland. Extract some additional functions for userland.