Add initial documentation for children component
Honestly, this is the area that still needs the most work. Handling
children is a little rough and requires a lot of manual wiring to get
working correctly. There is a lot of implementation knowledge required
to get children, especially iterated children, working correctly. Also
updated the `@@key` as an undocumented feature (which will work
if an object is emitted), but providing two separate methods for iterated
children is too confusing.

Ultimately, we need to replace all this noise with an actual templating
language that does all this for us.