Spike out silt, Handlebars-based templating
Only supports onMount to start, but gives us a base on which to build
the rest of the templating. Unfortunately, Handlebars templates are
compiled at runtime currently, but once we have an initial
implementation, we can use that to build a babel plugin on top of it to
extract the templates and precompile the tagged templates so the
runtime cost is eliminated.