Update `modifyChildProps` to take child `key`
The preferred method for mapping a parent's props to a child is to
fetch it in userland, rather than relying on framework magic to fetch
the correct key. This exposes the implementation to the user so it's
clear how the parent's props becomes the child. We should seek to
avoid as much magic in the framework as possible, and use functions
where possible to allow the user to implement it how she chooses.