Extract `child` factory function from children
All of the validation for an individual child definition can go into the
child factory function, which wraps the configuration into a standard
instance factory function. This reverses the order of the parameters
because we should be consistent across all of the instance factory
functions. Update the call sites to match this version.