Simplify the logic behind the stream unsubscribing
We don't need to hold onto the stream instances; we use `takeUntilBy`
to end the stream when the stream element is removed from the DOM.
Filters the mutations$ stream to only get the DOM mutations that are
relevant to the subscribe/unsubscribe streams. Updates the tests to
ensure the stream has been unsubscribed without needing to peek into
the `_currentSources` protected property. Instead, we just send an
event down to stream and see what comes out.