Refine main exports and define full module
The makes a number of changes in the exports of the framework:

1. rAF: the rAF-related actions move over the to the `actions` module,
and those, along with the `raf$` stream, get exported through the main
module. Fixes #26.
2. bootstrap: Removed. This is not the `start` function we're looking for.
We still plan on adding a `start` or `run` function, but `bootstrap` is not it.
Also removes the `BROOKJS_INIT` action. Userland should dispatch their
own actions on mount of they need to.
3. helpers: Exports all of the helpers individually rather than on a single
object. Better facilitates tree-shaking in consuming modules.
4. action: Previously unexported actions have been removed. If you need
them, they can be found [here][removed-actions]. These actions are too
generic to be usefully distributed with the framework.

  [removed-actions]: http://jamesdigioia.com/gistpens/removed-brookjs-actions/