Remove duplicate mutation actions from mutations$
As indicated in the test, if an element is added to the DOM, and a container
is added to that element in a separate transaction, the MutationObserver
emits two MutationRecords, one for each append. The previous logic of the
mutations$ didn't take this into account, which results in the NODE_ADDED
action being dispatched twice. This causes the children stream to mount
the node twice, which is incorrect behavior.