Override return value of diffhtml functions
This adds a task that returns an Observable, short-circuiting the
diffhtml task flow and returning the Observable to the consumer.
This makes diffhtml Observable-returning and an easy interop
with the rest of the framework. It also provides an initial entrance
into transitions, as userland could implement them as diffhtml
transition hooks, but they'd have to do them with Promises instead
of Observables (for now).