Update docs to remove `render$` from combinator
Since we've broken up the creation of the component stream into
two pieces (sink & source), we aren't going to provide all the
streams to combinator. Hooking into the render$ stream will be
provided through render-specific hooks, since it's fundamentally
a different process than the source streams.

This is a breaking change, but there isn't a lot one could do with
the render stream, because it doesn't emit any values.