Update rendering stream to be managed at the top
This updates the rendering stream to place the responsibility for
rendering in the hands of the top-level component. The domDelta
should mount that component, which updates the entire DOM in
response to the the emitted props.

`modifyChildProps` now really only gets passed into `onMount`,
which continues to provide a method for dropping down past the
framework to manipulate the DOM directly. To do so, blackbox
the component and configure the entire logic in `onMount`.

Otherwise, subcomponents should be configured the same way,
and the root component is response from assigning variables to
the Handlebars context as appropriate. Use `domDelta`'s
`selectProps` to map the store's normalized state to the view's
denormalized props. This makes it easier to map between the
BE and FE; though the back-end will have to do the same thing.

This also eliminates the context switching. Currently, if you're
using Handlebars on the BE (like in PHP), you end up thinking
through the entire template on page load, where the BE is
responsible for rendering, but on the FE, you have to tie
the `modifyChildProps` of the component to the individual
template of the child. With the root component managing the
whole thing, we can now just look at the Handlebars template as
the "source of truth" for the DOM.

This fundamentally means we're now streaming a single value into
the DOM, rather than plugging modified streams into subsections
of the DOM. I would not be surprised if this resulted in a perf
boost, although I haven't measured anything yet.

This solves, in my opinion, the biggest stumbling block for the
framework as a whole: the context switching between BE & FE.
Now, all those values can be transformed by 3 or 4 simple
functions, depending on how flexible the BE can be in structuring
props provided to the template.

BREAKING: Top-level component is responsible for rendering the
entire DOM. Method for blackboxing elements is now explicit.