Wrap events in ProxyEvents implementation
Borrowed from raynos/dom-delegator, the ProxyEvent wraps the
underlying DOM Event with a custom one, providing access to
the DOM nodes that represent both the container & the delegated
targets for the event.

You shouldn't see any impact on your applications, as all the
relevant properties are copied over. If any are missing, open an
issue and they can be provided (or use reflection if you must).

Fixes #17.