Document the bootstrapping process
I'm going to pass on building this into the library at this point.
What it would really be suggesting is the entire process of wiring
up the application should change. `run` should really take over
the entire application, and we'd just drop Redux integration.

Which would be weird, because we'd either need to deprecate
all of the Redux-specific modules, or maintain two ways of doing
everything. There isn't really a way of providing a `run` function
that uses Redux that is an improvement over just writing these few
lines in each application. Every application can thus bootstrap
itself however it likes, given its particular requirements. When we
add the CLI tool, it should use this in the generated files.