Remove current stub for animations and rendering
We don't want to document this as a "blessed" technique. Userland
should do whatever custom work they want in `onMount` and take
advantage of `renderFromHTML`, if their component requires it,
but animations and control of the rendering cycle will eventually be
part of brookjs proper, so there's no reason to encourage specific
alternative patterns.