Fix unit tests to be be template-driven
With `createElementFromTemplate`, we have a much easier way of
creating DOM fixtures to test against, so we can replace the complex
element-creating logic with templates. This allows us to break out the
test logic between templates rather than trying to shove it all into one.
We also don't have to manually create elements, which can be error-
prone.

Also fixes a bug when dealing with text nodes, which are being created
without their nodeValue and so need to be updated during the patch
process. Because we attempt to read the node before it's been created,
since a second patch is made for to update its value, we get an error
because it doesn't exist in the NodeCache yet. This doesn't feel right,
but it could be OK.

Also adds an Elements cache in Desalinate, to cleanup created elements
from diffHTML during unit tests.