Segregate, extend, and export Kefir
This does a few things for us:

1. It provides a single location where we can extend the built-ins from
the Kefir library while still being able to import & access as a whole.
2 It allows us to re-expot the library through the entry point so
consumers get access to the extensions we added. It also fixes an issue
I sometimes run into attempting to develop the library via `npm link`:
I get two instances of Kefir and the `instanceof` checks fail.
3. It provides a potential method for switching to a swappable
Observable implementation. The reference to `Kefir` could be replaced
with a getter that defers to an internally-registered library.
Feasibilty is unclear, but that's a far-future change.