Enable child streams to not use createSourceStream
In order to mount as template, we need to tell the children whether
to only return the source or the entire component. If the parent is
responsible for rendering its entire tree, then the children should
only return a source stream, but if it isn't, then the children should
manage their entire lifecyle.