Tell children to use factory if no parent render
If parent isn't passed a render function, it can still mount a section
of the DOM and then mount the child components, allowing them
to render themselves rather than be rendered by the parent. This
allows the parent to receive events from otherwise undynamic parts
of the DOM and take advantage of the architecture without needing
to buy into the particular rendering details.