Enable root blackboxed components to be rendered
A component at the root of the render tree is in control of its entire
tree, so we expect the component to be using `renderFromHTML` and
updating it as part of a custom render cycle. A component won't
blackbox itself _from itself_, but will want to blackbox itself from its
parent components. This ensures it can render itself as it chooses.

Comes with tests!