Move babel-register call inside `index.js`
Eventually, we'd like to swap out `babel-register` for `@std/emd`
but it's still early days for that package, so we're going to stick with
`babel-register` for now. This also makes it possible to, in the future,
point the `pkg.json#main` at `main.(m)js` when node gets native
ESM support.

Alternatively, if we decide we want to use Flowtype here, we'll need
babel anyway to strip the types and add type-checking, which would
also be a viable option from here.

Remove the unambiguous grammer rule, because these entry points
should be scripts.