Switch compilation to @std/esm
`@std/esm@0.6.0` resolves the last problems with the package,
and we're able to properly use it as the bootstrapping of the cli.
It's not yet working with `ava`, but that's less important at this
point. The goal is to get to a no-compile, fully-ES6 CLI app.
That does preclude the use of Flow, but we'd then introduce a
build step with `babel-cli` and precompile it, rather than using
`babel-register` (as we are now). And even still, we should
eventually drop it in the tests.

Hopefully, this also makes `npx` work for `beaver`.