Add name to `make` command + lenses
We're not going to prompt the user after they type the
command. We should get that information when the user runs
the command in the first place and just take it from there. This
is much simplified by comparison. We can prompt them and
make it optional, which would give us a place to provide some
info to the user about best practices & naming conventions.