Normalize tuples to 3 elements
While writing the types for the AST, I found Flow would complain
about trying to create a Union type out of multiple tuples of
different lengths. This includes 3 elements in every tuple,
roughly corresponding to type, attributes, and children. In text
nodes and silt expressions, this includes adding an extra element
to make it the same as all the others. This will make writing
the types easier while maintaining the compact serialization
provided by the tuple types here.