Move constants to types and add @flow
We've got one place to fix, because we do need to add a check
to make sure the block type is a valid type, but I'm not sure how
to get Flow to do that without manually checking all the types.
Maybe getting flow-runtime up in here to provide those checks
would work?