Add tests for createElementFromTemplate
We need to run the tests from the browser only because we can't
get these working in node & the browser at the same time yet.
We'll have to put together a test runner that inject the jsdom
globals into the environment before executing the tests. That will
need to be build into brookjs-cli's test runner, which we could then
use... to run all our tests.