Add mocha + nyc on cli w jsdom
These tests don't pass yet because we don't have a MutationObserver
in jsdom. We have to modify the children mounting logic to use a
diffhtml middleware to intercept appended children elements. Only
after we enable that will we be able to run the unit tests in node.

Also initially integrate chai-kefir. We might be able to eliminate our
usage of `sinon`. We're going to overhaul the whole testing process
to allow us to run the tests in node. We'll eventually be able to use
this to handle tests in userland.