Remove desalinate internal package from brookjs
Depend on external package instead. Weird path in render is
because there's a circular dep, which isn't itself a problem but
webpack seems to be resolving brookjs from two different
places, `src` in the test files and `es` when importing from
`brookjs` in `desalinate`. This points them both to `es`. This
also suggests we may need to both break this up into more
smaller packages and move the testing to a `test` folder like
`kefir-chai` does and treat the component and its subdeps
from the outside only. If we start doing this is a more legit BDD
fashion, we can probably write better tests of the `component`
feature if we extract it to a separate package.