Update render process to use stream of `effect$`s
Instead of relying on a MutationObserver to bind children to
their parent, the root component (responsible for rendering the
entire DOM), generates a stream of effects, which it passes
down to the children. This stream is then threaded down to the
`children` function, which uses it to bind child factories to their
parent streams. We can later use this stream to thread through
the animations callback to control when renders occur.

Now that we no longer rely on MutationObserver, we can now
run the tests in Node w/ jsdom, falling back to Karma for browser
verification. Those tests have been added to the npm test
command.