Add support for webpack config in beaverrc
This allows the user to provide values relative to a standard
brookjs app's other configuration. Output is still relative to
the `cwd` but the entry configuration is relative to the
app's `dir` config. Lastly, `modifer` provides access to the
webpack cofiguration before it gets passed to webpack,
so the user can configure webpack however they want.