Switch brookjs-cli to use babel instead of @std/esm
We're going to need a build step anyway, to convert the JSX
from Ink, so we might as well have the build step handle the
modules as well. Use our new babel-preset, as well as flow
here. We're going to drop lenses and maybe try and drop
ramda completely. We're maintaining @std/esm to load up the
`.beaverc.js` file for now.