Fix bug in Collector with Observable children
This _feels_ like a bug that _should_ be solved in `h`, but I can't
for the life of me figure out why it's working this way. It makes
me a bit nervous because any outside package that messes with
the `children` like Collector does will break. This should be
resolved before this hits `0.12.0`.