Fix integration with Collector in `h`
We're adding the prop to lift components into event emitting
streams. This is used by the `walkChildren` function to decide
if and how we should wrap the elements. We use some prtty
hacky implementations, but this seems to work for now. Not
sure creating a new class and extending each time is the best
approach, but there's definitely some optimizations to be
made here.