Add implementation of `loop` as iteration helper
Takes an `Observable<{ order, dict }>`, which it uses to build up the
correct Element + stream combination. It also comes with a
built-in optimization, preventing new elements from being emitted if
the order doesn't change, and reusing the streams created. This way,
the reconciler can treat the returned element as the same thing,
avoid unncessary rerenders.

Fixes #124.