Replace Reffable with withRef$
We're not going to be able to go the lifecycle route, I don't think.
If we do, we may want to consider a HOC that enables that as well,
but for now, withRef$ works.

The tests are disabled because Enzyme doesn't currently support
React 16.3, unfortunately. Will re-enable them once that gets fixed.