Move to rollup and hoist configuration
All of the dev dependencies live in the root package.json now, so that's
been added to greenkeeper.json. The package.json's have been updated
to use beaver's dev tdd command to watch and rerun tests as we need.
We're using main / module / browser by rollup to generate the various
bundles, so this ensures we have a relationship between those two.
All the webpack configurations get deleted. The build scripts are
simplified to use rollup and the root rollup config. Some minor adjustments
were required to imports to make them play nice w/ rollup.
We also simplified the babelrc, as rollup is handling outputting
cjs / esm / umd. We're using the older version of babel-plugin-ramda
because of this issue:

https://github.com/megawac/babel-plugin-ramda/issues/41

We decided to switch rollup now because webpack appears to be
increasingly geared towards apps. The initial switch has it outputting
a main.js with the webpack runtime, separate from the main bundle.
Since we needed to switch to rollup at some point anyway, this seemed
like a good time to do it. And hoisting the new config allows us to simplify
the overall development process, so we get a bunch of benefits at once.

Fixes #52. Fixes #181.