Add eddy module for returning actions from reducer
This gives us the ability to centralize all our business logic in the
store, including not only what state we're in now but what happens
next. This allows us to decouple our deltas from the source of the
actions, instead relying on the store to manage data flow through
the system.

Fixes #152.