Add test for interleaving behavior
Here's the reason this emits in this order:

When FIRST_REQUEST is emitted, two actions are emitted (and queued) in response:
SECOND_REQUEST & FIRST_RESPONSE. The first action is emitted, and in response,
SECOND_RESPONSE is emitted. Once SECOND_REQUEST completes its pass through
the middleware, it drains its queue before yielding the callstack, emitting
SECOND_RESPONSE. After it does, the parent queue continues, emitting
FIRST_RESPONSE. That's how we result in emitting that order.