Initial implementation of Storybook integration
Piggyback mostly off `addons-actions`, but sets up the API we're going to be
setting in place. We need both the decorator & registration. The decorator
wraps every story in a RootJunction and sends them to the Manager. That
is responsible for then rendering the emitted actions in Storybook.

We just link our decorator to the panel rendered by addons-actions, but
we can swap that out for our own custom display.