Upgrade brookjs-desalinate to TypeScript
Add some initial types for dependencies we have, added to the tsconfig.json.
The rollup configuration was cleaned up to get the build working correctly,
only compiling unpkg if we have that set. We're not going to set it for
desalinate because we're having issues with that build. We can readd this
later if we want. The types for the RootJunction were improved, and the
types were adding to silt's pkg.json. Also update the build scripts to
generate the declaration files outside of rollup. We're not generating a
umd build for desalinate, so we gotta do it outside.