Collect new app information from the user
Implement the cli UI to prompt the user with a series of
questions and save the information to the store. We also
needed to switch off beaver's testing tool, so the command
is back to standard mocha, but we were able to write
unit tests to cover the entirety of the View. We still need
to implement the "complete" view, but we're going to handle
the `exec` method first and design the UI based on the
results.

We added types for various packages that we would like to
upstream as well. We also need to figure out the issue with
the View component as a member of the NewCommand
class. The types aren't lining up and I'm getting strange
errors. It could be the `Configured/ConfiguringState` types,
which is overcomplicating things but keeps use from
needing to `null`-check everything as we proceed through
the process.