Finish initial new command wiring
Continue to defer most of the internal logic to hygen to
generate the initial app scaffolding. Clean out some dead
code we repurposed in the app. We're also going to add a
Services generic to the base Command, as we're going to
want to inject those services for testing, although we haven't
wired them up yet.