Fix feature testing setup
Switch from cp.spawn => node-pty so it properly emulates
a TTY terminal, allowing `setRawMode` to work as expected.
Prettier was run over the support files, but the only thing
that changed was the spawn handling as well as how we
send typing to the child process. We do add a small wait
between each keypress, as we got an error from "typing"
too fast.