Wrap babel-preset-brookjs around CRA's preset
I'm not really sure what, if anything, I want to do with this yet.
CRA's preset is _really good_, and we're probably better off
just using theirs. We can use it to add stuff we want later, if we
find it's missing something (or it's taking too long to add new
proposals), but this is ultimately not that helpful over the original.
It does give us something to standardize some internal usages
around though.

We can also set up the babel registering once, and save the
areas we want it to actually babel as an array.