Lift Webpack configuration from CRA
Some things aren't quite working yet, and we made a few
adjustments for things we prefer, but overall, pretty closely
matches their approach.

The strangest thing is rollup complaining about react-dev-utils
and some of the deep imports we're doing into it. Not sure why
but will want to get that fixed.

Fixes #406.