Extract brookjs-eddy to a separate package (#952)
This package is all of the redux-specific tooling, including a new
`createStore` which solves the weird eddy typing issue.
Unfortunately, this isn't really solvable for `eddy` directly because
no enhancer can change the type of the `reducer`. If you're not
using TypeScript, `eddy` works fine as-is. If you are, you can
use `createStore`. If you need more than what `createStore`
provides, you'll either have to use `eddy` weirdly or cast your
`reducer` to a regular redux reducer.

This also simplifies the app, as you only need to install the root
`brookjs` package, and use its `createStore`, and completely
ignore the subpackages. This also required a fix in the rollup
config to ensure imported subpackages are correctly ignored.
This was an issue caused by `react-devtools-extension`
explicitly, as it caused `redux` to be default imported, but it's
also an issue because more packages are being bundled into
the final package, including some `react-dev-utils` code as well
as the `@babel/runtime` helpers.

Fixes #380.